Supress or Resolve compile-time warnings
During build with full warnings, several items were flagged by the
compiler. Most warning indicators involved conversion to/from QVariant
and a locally defined enumerated type. Issues of this type were resovled
with explicit type casting to as required by called function.

There was also needed a few explicit casting conversions to integer type
required in code to supress type conversion warnings( size_t to int).

Defined value PROGRESS_MAX was changed to an integer type (remove '.0').
This resolved warnings where PROGRESS_MAX was apart of arithmetic operations
involving other integer typed values and warnings of their conversion
to float/double.

+++ gui/checkthread.cpp
@@ -197,7 +197,7 @@ void CheckThread::runAddonsAndTools(const ImportProject::FileSettings *fileSetti
                 QFile f1(analyzerInfoFile + '.' + addon + "-E");
                 if (f1.open(QIODevice::ReadOnly | QIODevice::Text)) {
                     QTextStream in1(&f1);
-                    const quint16 oldchksum = in1.readAll().toInt();
+                    const quint16 oldchksum = static_cast<quint16>(in1.readAll().toUInt());
                     if (oldchksum == chksum) {
                         QFile f2(analyzerInfoFile + '.' + addon + "-results");
                         if (f2.open(QIODevice::ReadOnly | QIODevice::Text)) {
